+++ examples/init-update-contract.rs
@@ -1,4 +1,3 @@
-#![allow(deprecated)]
 //! Basic example that shows how to initialize and update a smart contract.
 //!
 //! In particular, it uses the "weather" contract which is part of the
@@ -14,9 +13,10 @@ use concordium_rust_sdk::{
     },
     types::{
         smart_contracts::{ModuleReference, OwnedParameter},
-        transactions::{send, BlockItem, InitContractPayload, UpdateContractPayload},
+        transactions::{send, InitContractPayload, UpdateContractPayload},
         AccountInfo, WalletAccount,
     },
+    v2::{self, BlockIdentifier},
 };
 use std::path::PathBuf;
 use structopt::*;
@@ -27,7 +27,7 @@ struct App {
     #[structopt(
         long = "node",
         help = "GRPC interface of the node.",
-        default_value = "http://localhost:10000"
+        default_value = "http://localhost:20000"
     )]
     endpoint:  endpoints::Endpoint,
     #[structopt(long = "account", help = "Path to the account key file.")]
@@ -89,17 +89,17 @@ async fn main() -> anyhow::Result<()> {
         App::from_clap(&matches)
     };
 
-    let mut client = endpoints::Client::connect(app.endpoint, "rpcadmin").await?;
+    let mut client = v2::Client::new(app.endpoint).await?;
 
     // load account keys and sender address from a file
     let keys: WalletAccount =
         WalletAccount::from_json_file(app.keys_path).context("Could not parse the keys file.")?;
 
-    let consensus_info = client.get_consensus_status().await?;
     // Get the initial nonce at the last finalized block.
     let acc_info: AccountInfo = client
-        .get_account_info(&keys.address, &consensus_info.last_finalized_block)
-        .await?;
+        .get_account_info(&keys.address.into(), BlockIdentifier::LastFinal)
+        .await?
+        .response;
 
     let nonce = acc_info.account_nonce;
     // set expiry to now + 5min
@@ -136,9 +136,8 @@ async fn main() -> anyhow::Result<()> {
         }
     };
 
-    let item = BlockItem::AccountTransaction(tx);
     // submit the transaction to the chain
-    let transaction_hash = client.send_block_item(&item).await?;
+    let transaction_hash = client.send_account_transaction(tx).await?;
     println!(
         "Transaction {} submitted (nonce = {}).",
         transaction_hash, nonce,