+++ examples/send-transfer.rs
@@ -1,15 +1,14 @@
-#![allow(deprecated)]
 //! Basic example that shows how to send a transaction, in this case a transfer
 //! from the account to itself.
 use anyhow::Context;
 use clap::AppSettings;
 use concordium_rust_sdk::{
     common::types::{Amount, TransactionTime},
-    endpoints,
     types::{
         transactions::{send, BlockItem},
-        AccountInfo, WalletAccount,
+        WalletAccount,
     },
+    v2::{self},
 };
 use std::path::PathBuf;
 use structopt::*;
@@ -19,9 +18,9 @@ struct App {
     #[structopt(
         long = "node",
         help = "GRPC interface of the node.",
-        default_value = "http://localhost:10000"
+        default_value = "http://localhost:20000"
     )]
-    endpoint:  endpoints::Endpoint,
+    endpoint:  v2::Endpoint,
     #[structopt(long = "account", help = "Path to the account key file.")]
     keys_path: PathBuf,
 }
@@ -33,19 +32,18 @@ async fn main() -> anyhow::Result<()> {
         let matches = app.get_matches();
         App::from_clap(&matches)
     };
-    let mut client = endpoints::Client::connect(app.endpoint, "rpcadmin").await?;
+    let mut client = v2::Client::new(app.endpoint).await?;
 
     // load account keys and sender address from a file
     let keys: WalletAccount =
         WalletAccount::from_json_file(app.keys_path).context("Could not read the keys file.")?;
 
-    let consensus_info = client.get_consensus_status().await?;
     // Get the initial nonce at the last finalized block.
-    let acc_info: AccountInfo = client
-        .get_account_info(&keys.address, &consensus_info.last_finalized_block)
+    let nonce = client
+        .get_next_account_sequence_number(&keys.address)
         .await?;
 
-    let nonce = acc_info.account_nonce;
+    let nonce = nonce.nonce;
     // set expiry to now + 5min
     let expiry: TransactionTime =
         TransactionTime::from_seconds((chrono::Utc::now().timestamp() + 300) as u64);