+++ examples/total-stake.rs
@@ -1,11 +1,6 @@
-#![allow(deprecated)]
 //! List total equity, total delegated stake, and effective stake.
 use clap::AppSettings;
-use concordium_rust_sdk::{
-    common::types::Amount,
-    endpoints::{self, Endpoint},
-    types::{hashes::BlockHash, BakerPoolStatus, PoolStatus},
-};
+use concordium_rust_sdk::{common::types::Amount, v2, v2::BlockIdentifier};
 use structopt::StructOpt;
 
 #[derive(StructOpt)]
@@ -13,16 +8,15 @@ struct App {
     #[structopt(
         long = "node",
         help = "GRPC interface of the node.",
-        default_value = "http://localhost:10000"
+        default_value = "http://localhost:20000"
     )]
-    endpoint: Endpoint,
+    endpoint: v2::Endpoint,
     #[structopt(
         long = "block",
+        default_value = "lastfinal",
         help = "Block to query the data in. Defaults to last finalized block."
     )]
-    block:    Option<BlockHash>,
-    #[structopt(long = "token", help = "GRPC login token", default_value = "rpcadmin")]
-    token:    String,
+    block:    BlockIdentifier,
 }
 
 #[tokio::main(flavor = "multi_thread")]
@@ -33,29 +27,19 @@ async fn main() -> anyhow::Result<()> {
         App::from_clap(&matches)
     };
 
-    let mut client = endpoints::Client::connect(app.endpoint, app.token).await?;
-
-    let consensus_info = client.get_consensus_status().await?;
-
-    let block = app.block.unwrap_or(consensus_info.last_finalized_block);
+    let mut client = v2::Client::new(app.endpoint).await?;
 
     let mut equity: u64 = 0;
     let mut delegated: u64 = 0;
     let mut effective: u64 = 0;
 
-    let birk_params = client.get_birk_parameters(&block).await?;
+    let birk_params = client.get_election_info(app.block).await?;
+    let block = birk_params.block_hash;
 
     let mut active_bakers: u32 = 0;
-    for baker in birk_params.bakers {
-        let pool = client.get_pool_status(Some(baker.baker_id), &block).await?;
-        if let PoolStatus::BakerPool {
-            status:
-                BakerPoolStatus {
-                    current_payday_status: Some(current_payday_status),
-                    ..
-                },
-        } = pool
-        {
+    for baker in birk_params.response.bakers {
+        let pool = client.get_pool_info(&block, baker.baker_id).await?.response;
+        if let Some(current_payday_status) = pool.current_payday_status {
             active_bakers += 1;
             equity += current_payday_status.baker_equity_capital.micro_ccd();
             delegated += current_payday_status.delegated_capital.micro_ccd();