+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Unreleased changes
 
+- Remove the V1 API.
+- Add `Display` implementation to `BlockIdentifier`.
+- Add `Display` and `FromStr` implementations for `AccountIdentifier`.
+- Rename `find_first_finalized_block_no_later_than` into
+  `find_first_finalized_block_no_earlier_than` since that correctly reflects its
+  semantics with respect to time and is much clearer.a
+
 ## 3.2.0
 
 - The sdk now requires a `rustc` version at least 1.67 (Before it required version 1.66).